--- Remember/ICalendario.java
@@ -1,7 +1,0 @@
-package Remember;
-
-public interface ICalendario {
-    int getDia();
-    int getMes();
-    int getAno();
-}
