--- Remember/Observer.java
@@ -1,7 +1,0 @@
-package Remember;
-
-public interface Observer  {
-	
-	void atualizar(Double novoinicio);
-
-}
