+++ Remember/Tarefa/FabricaDeTarefas.java
@@ -2,8 +2,16 @@ package Remember.Tarefa;
 
 import Remember.Componente;
 
-public interface FabricaDeTarefas {
-    
-    Componente getTarefa(int e);
+public class FabricaDeTarefas implements InterfaceFabrica {
+   
+    @Override
+	public Componente getTarefa(int e) {
 
+		if(e == 1)
+			return new TarefaSimples();
+		
+		else if(e == 2)
+			return new TarefaCustomizavel();
+		return null;
+	}
 }
