+++ Remember/Tarefa/InterfaceFabrica.java
@@ -1,0 +1,9 @@
+package Remember.Tarefa;
+
+import Remember.Componente;
+
+public interface InterfaceFabrica {
+    
+    Componente getTarefa(int e);
+
+}
